Add files via upload
This is a basic front-end web page clone of Spotify, created for learning and practice purposes. It includes a simplified layout resembling Spotify's main page, with sections for music playlists, recommended artists, and a search bar. This project is ideal for beginners looking to explore HTML, CSS, and JavaScript for web development.